separate files into different folders
